+++ src/vk_initializers.h
@@ -116,6 +116,6 @@ namespace vkinit {
 
     void initMesh(Mesh& mesh, VkDevice device, VkCommandPool commandPool, VkQueue queue, VmaAllocator allocator,
                   const std::vector<Vertex>& vertices,
-                  const std::vector<uint16_t>& indices);
+                  const std::vector<uint32_t>& indices);
 
 }
